refactor CouponUi : 쿠폰 도메인 모델을 ui 모델로 변경하는 역할
- before: 도메인 모델에서 ui 모델로 매핑
- after: ui 모델에서 도메인 모델로 매핑
