+++ app/src/main/java/woowacourse/shopping/domain/model/Coupon.kt
@@ -1,6 +1,5 @@
 package woowacourse.shopping.domain.model
 
-import woowacourse.shopping.ui.model.CouponUi
 import java.time.LocalDate
 
 sealed class Coupon {
@@ -14,38 +13,3 @@ sealed class Coupon {
 
     abstract fun discountAmount(orders: Orders): Int
 }
-
-fun Coupon.toUi(): CouponUi =
-    when (this) {
-        is FixedAmountCoupon ->
-            CouponUi(
-                id = id,
-                description = description,
-                expirationDate = expirationDate,
-                minimumAmount = minimumAmount,
-            )
-
-        is FreeShippingCoupon ->
-            CouponUi(
-                id = id,
-                description = description,
-                expirationDate = expirationDate,
-                minimumAmount = minimumAmount,
-            )
-
-        is BoGoCoupon ->
-            CouponUi(
-                id = id,
-                description = description,
-                expirationDate = expirationDate,
-                minimumAmount = null,
-            )
-
-        is PercentageCoupon ->
-            CouponUi(
-                id = id,
-                description = description,
-                expirationDate = expirationDate,
-                minimumAmount = null,
-            )
-    }
